# backend/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lynx Server API
  version: 0.1.0
  description: Local Lynx Guardian backend API served by the Gin-based lynx-server binary.
servers:
  - url: http://127.0.0.1:31789
    description: Default local lynx-server
...
paths:
...
  /lynx/approvals/{approvalId}/resolve:
    post:
      tags: [query]
      operationId: resolveApproval
      summary: Resolve one approval record.
      description: Allows a still-pending non-L4 approval from the local console helper view and records the approval result. L4 hard-deny records cannot be approved from this endpoint.
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalResolveRequest"
      responses:
        "200":
          $ref: "#/components/responses/ObjectResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    ObjectResponse:
      description: Single JSON object result.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JsonObject"
    ErrorResponse:
      description: Error payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
          enum: [false]
        message:
          type: string
    ApprovalResolveRequest:
      type: object
      properties:
        approvalId:
          type: string
        resolution:
          type: string
          enum: [allow-current-chain, allow-once, allow-always, approved, rejected, denied, blocked, failed]
        chainId:
          type: string
        sessionKey:
          type: string
        channelProfile:
          type: string
        channelId:
          type: string
        conversationId:
          type: string
        requesterId:
          type: string
        requesterOuId:
          type: string
        approverId:
          type: string
        approverOuId:
          type: string
        riskFamily:
          type: string
        riskLevel:
          type: string
          enum: [L0, L1, L2, L3, L4]
        toolName:
          type: string
        targetKind:
          type: string
        targetHash:
          type: string
        resourceScope:
          $ref: "#/components/schemas/JsonObject"
        expiresAt:
          type: string
      additionalProperties: true
...